moved GlobalEvent location to components folder (#303)
