+++ src/components/stepper/stepper.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import {SBComponent} from 'stylable-react-component';
-import GlobalEvent from '../../common/global-event';
 import {ChevronDownIcon, ChevronUpIcon} from '../../icons';
 import buttonStyles from '../../style/default-theme/controls/button.st.css';
+import {GlobalEvent} from '../global-event';
 import styles from './stepper.st.css';
 
 export interface StepperProps extends React.HTMLProps<HTMLElement> {
